Conform and fix SGS SD card stuff inside Android OS (1/4)
Mount internal storage to emmc (like other devices), but the expected behavior for
new users is for it to be mounted to /sdcard (thus eliminating the requirement of
an external SD card). This will be fixed in the device build files by defaulting
to having the switch enabled.

Patchset 2: This is to squash a "bug-but-not-a-bug" issue. Technically the partition
mapping was wrong, but it snuck through and worked anyway as they both happen to "match"
under default circumstances. Best to make sure and get it right though.

Change-Id: I38b87cd8d20f66b7ee76ddc240ab1c98954a7c25